==> 2-Spotipy_Categories.ipynb <==
### Objective

Using spotipy API to retrieve songs from various playlists.  
**First step is retrieving all categories and their id's.**

In order to not send too many API request of same nature, I am saving the intermediate result. 

*Output: 2_Category_Ids.csv*, columns 'category_id', 'category_name'

import spotipy
from spotipy.oauth2 import SpotifyClientCredentials
import time
import pandas as pd
from spotipy_util import retrieve_sp

> connect to api

sp = retrieve_sp()

> retrieve available categories

all_category_ids = []

# Define the total number of categories you want to retrieve
total_categories = 100
categories_per_request = 50

# Iterate over multiple requests until the desired number of categories is reached
for offset in range(0, total_categories, categories_per_request):
    # Get categories with the current offset
    categories = sp.categories(country=None, locale=None, limit=categories_per_request, offset=offset)
    
    # Extract category IDs and add them to the list
    category_info = [(category['id'], category['name']) for category in categories['categories']['items']]
    all_category_ids.extend(category_info)

# Print the list of all category IDs
len(all_category_ids)

all_category_ids

> save to csv, convert the list of tuples to a DataFrame

df = pd.DataFrame(all_category_ids, columns=['category_id', 'category_name'])
df.to_csv('./cache/2_Category_Ids.csv', index=False)

df
