==> 3-Spotipy_Playlists.ipynb <==
### Objective

Using spotipy API to retrieve songs from various playlists.  
**Second step is retrieving playlists and their id's for each category previously retrieved.**

In order to not send too many API request of same nature, I am saving the intermediate result. 

*Output: 3_Playlist_Ids.csv*, columns 'category_id', 'category_name', 'playlist_id', 'playlist_name'

import pandas as pd
from spotipy_util import retrieve_sp

> retrieve categories

categories = pd.read_csv('./cache/2_Category_Ids.csv')
display(categories.head(1))
display(categories.shape)

> connect to api

sp = retrieve_sp()

> there is a total off 55 categories  
> retrieve up to 5 playlists per category

all_playlist_data = []

# Iterate over each category
for idx, row in categories.iterrows():
    category_id = row['category_id']
    category_name = row['category_name']
    
    # Get playlists for the current category
    playlists = sp.category_playlists(category_id, country=None, limit=5, offset=0)
    
    # Iterate over each playlist and append its data to the list
    for playlist in playlists['playlists']['items']:
        playlist_id = playlist['id']
        playlist_name = playlist['name']
        all_playlist_data.append({'category_id': category_id, 'category_name': category_name, 'playlist_id': playlist_id, 'playlist_name': playlist_name}) 


# sanity check 
# len(all_playlist_data)
# all_playlist_data

# Convert the list of playlist IDs to a DataFrame
playlist_df = pd.DataFrame(all_playlist_data)

playlist_df

import pandas as pd
playlist_df = pd.read_csv('./cache/3_Playlist_Ids.csv')


playlist_df['category_name'].unique()

playlist_df = playlist_df[~playlist_df['category_name'].isin(['Made For You', 'New Releases', 'K-pop', 'Anime', 'Nature & Noise', 'Spotify Singles', 'Tastemakers', 'Netflix'])]


playlist_df['category_name'].unique()

> save to csv, convert the list of tuples to a DataFrame

playlist_df.to_csv('./cache/3_Playlist_Ids.csv', index=False)
