==> 4-Spotipy_Songs.ipynb <==
### Objective

Using spotipy API to retrieve songs from various playlists.  
**Third step is retrieving the song id's from the playlists.**

In order to not send too many API request of same nature, I am saving the intermediate result. 

*Output: 4_Song_Ids.csv*, columns 'category_id', 'category_name', 'playlist_id', 'playlist_name', 'song_id', 'song_artist', song_title

import pandas as pd
from spotipy_util import retrieve_sp
import time

> retrieve categories

playlist_df = pd.read_csv('./cache/3_Playlist_Ids.csv')
display(playlist_df.head(1))
display(playlist_df.shape)

playlist_df['category_name'].unique()

> connect to api

sp = retrieve_sp()

> retrieve all song id's, artists, titles from each playlist

all_songs = []

for idx, row in playlist_df.iterrows():
    playlist_id = row['playlist_id']
    category_id = row['category_id']
    category_name = row['category_name']
    
    # Get playlist details
    playlist = sp.playlist(playlist_id)
    playlist_name = playlist['name']
    
    # Get tracks for the current playlist
    tracks = sp.playlist_tracks(playlist_id, fields='items.track.id,items.track.artists,items.track.name,total')

    # Check if tracks is not None and 'items' in tracks
    if tracks is not None and 'items' in tracks:
        # Iterate over each track and append its data to the list
        for item in tracks['items']:
            # Check if 'track' key exists in the item and it's not None
            if 'track' in item and item['track'] is not None:
                # Check if 'id' key exists in the track
                if 'id' in item['track']:
                    song_id = item['track']['id']
                    song_title = item['track']['name']
                    
                    # Join artists' names if they exist
                    if 'artists' in item['track'] and item['track']['artists']:
                        song_artist = ', '.join([artist['name'] for artist in item['track']['artists'] if 'name' in artist])
                    else:
                        song_artist = None  # or any other appropriate placeholder
                    
                    all_songs.append({'category_id': category_id, 'category_name': category_name, 'playlist_id': playlist_id, 'playlist_name': playlist_name, 'song_id': song_id, 'song_artist': song_artist, 'song_title': song_title})

    # Introduce a delay between requests, increasing with each iteration
    time.sleep(2)

> check, remove duplicates

songs_df = pd.DataFrame(all_songs)
songs_df.shape

songs_df.drop_duplicates(inplace=True) 
songs_df.shape   

songs_df

> save to csv, convert the list of tuples to a DataFrame

songs_df.to_csv('./cache/4_Song_Ids.csv', index=False)
