==> 5-Spotipy_Audio_Features.ipynb <==
### Objective

Using spotipy API to retrieve songs from various playlists.  
**Fourth step is retrieving the songs audio features.**

In order to not send too many API request of same nature, I am saving the intermediate result. 

*Output: 5_Song_Audio_Features.csv*, columns 'category_id', 'category_name', 'playlist_id', 'playlist_name', 'song_id', 'song_artist', song_title

import pandas as pd
from spotipy_util import retrieve_sp
from time import sleep
from random import randint
from pandas import json_normalize

> retrieve songs

audio_df = pd.read_csv('./cache/4_Song_Ids.csv')
audio_df.columns

> connect to api

sp = retrieve_sp()

chunks = [(i, i+100) for i in range(0, len(audio_df), 100)]
audio_features_list = []
for chunk in chunks:
    id_list100 = audio_df['song_id'][chunk[0]:chunk[1]]
    audio_features_list = audio_features_list + sp.audio_features(id_list100)
    sleep(randint(1,3000)/1000)
len(audio_features_list)

len(audio_features_list)

audio_features_df = json_normalize(audio_features_list)
audio_features_df.tail()

audio_features_df.shape

> merge in the merged_df to have it all in one

# Drop duplicates in audio_df based on the 'id' column
audio_features_df.drop_duplicates(subset=['id'], inplace=True)

# Drop duplicates in audio_features_df based on the 'song_id' column
audio_df.drop_duplicates(subset=['song_id'], inplace=True)

audio_features_df.shape
audio_df

merged_df = pd.merge(audio_df, audio_features_df, left_on='song_id', right_on='id', how='left')
merged_df.dropna(inplace=True)

merged_df.isna().sum().sum()

> save to csv, convert the list of tuples to a DataFrame

merged_df.to_csv('./cache/5_Song_Audio_Features.csv', index=False)
