==> 7-Recommender.ipynb <==
### Objective

**Create a recommender that gives a song recommendation based on a user given input.**

Model Logic:

1. **Check if the Song is in the Hot 100 List:**
   - If the song is in the Hot 100 list: *(1_Top_100.csv)*
     - Respond with another song from the Hot 100.
   - If the song is not in the Hot 100 list:
     - Proceed to step 2.

2. **Find Audio Features of the Song Using the Spotify API:**
   - Utilize the Spotify API to obtain audio features (e.g., danceability, energy, tempo) of the song.

3. **Apply Cluster Model to the Song:**
   - Apply a pre-trained cluster model to the song's audio features. *(cluster_model.pkl - and scaler.pkl for scaling)*
   - Determine the cluster to which the song belongs.

4. **Recommend a Song from the Same Cluster:**
   - Retrieve a song from the same cluster as the given song.
   - This recommendation can be based on similarity in audio features or other criteria.


import pickle
import numpy as np
import pandas as pd

from sklearn.cluster import KMeans
from recommender import song_recommender, is_hot_100, get_audio_features, determine_cluster, recommend_song_from_cluster

> retrieve data

spotify = pd.read_csv('cache/6_clustered_df.csv')
top_100 = pd.read_csv('cache/1_Top_100.csv')
top_100

> implement recommender

# in top 100, artist 
user_input = 'rm'
if is_hot_100(user_input, top_100):
    other_song = top_100[top_100['artist'].str.lower() != user_input].sample(n=1).iloc[0]
    print(f"You might like: '{other_song['artist']}' - '{other_song['title']}'")
    print('(in top 100)')
else:
    # get audio feature of the input song 
    audio_features = get_audio_features(user_input)
    # Apply cluster model and recommend a song from the same cluster
    cluster = determine_cluster(audio_features)
    recommended_song = recommend_song_from_cluster(spotify, cluster)
    print(f"You might like: '{recommended_song['song_artist'].iloc[0]}' - '{recommended_song['song_title'].iloc[0]}'")
    print('(from Spotify)')

# in top 100, title
user_input = 'Come back to me'
if is_hot_100(user_input, top_100):
    other_song = top_100[top_100['artist'].str.lower() != user_input].sample(n=1).iloc[0]
    print(f"You might like: '{other_song['artist']}' - '{other_song['title']}'")
    print('(in top 100)')
else:
    # get audio feature of the input song 
    audio_features = get_audio_features(user_input)
    # Apply cluster model and recommend a song from the same cluster
    cluster = determine_cluster(audio_features)
    recommended_song = recommend_song_from_cluster(spotify, cluster)
    print(f"You might like: '{recommended_song['song_artist'].iloc[0]}' - '{recommended_song['song_title'].iloc[0]}'")
    print('(from Spotify)')

# not in top 100
user_input = 'The Code'
if is_hot_100(user_input, top_100):
    other_song = top_100[top_100['artist'].str.lower() != user_input].sample(n=1).iloc[0]
    print(f"You might like: '{other_song['artist']}' - '{other_song['title']}'")
    print('(in top 100)')
else:
    # get audio feature of the input song 
    audio_features = get_audio_features(user_input)
    # Apply cluster model and recommend a song from the same cluster
    cluster = determine_cluster(audio_features)
    recommended_song = recommend_song_from_cluster(spotify, cluster)
    print(f"You might like: '{recommended_song['song_artist'].iloc[0]}' - '{recommended_song['song_title'].iloc[0]}'")
    print('(from Spotify)') 

> live usage

#song_recommender(top_100, spotify)
